Move shared components into the componets folder
