+++ src/components/ProgramList.js
@@ -11,9 +11,7 @@ import Grid from "@material-ui/core/Grid";
 import Paper from "@material-ui/core/Paper";
 import Box from "@material-ui/core/Box";
 
-import Placeholder from "./static/images/image-placeholder.png";
-
-const IMAGE_URL = "https://image.tmdb.org/t/p/w780/";
+import { getImageURL } from "./ProgramDetails";
 
 const useStyles = makeStyles((theme) => ({
   card: {
@@ -30,16 +28,9 @@ const useStyles = makeStyles((theme) => ({
   sectionTitle: { marginBottom: theme.spacing(2) },
 }));
 
-function PopularList({ items, listTitle }) {
+function ProgramList({ items, listTitle }) {
   const classes = useStyles();
 
-  const getImageURL = (imagePath) => {
-    if (imagePath) {
-      return `${IMAGE_URL}${imagePath}`;
-    } else {
-      return Placeholder;
-    }
-  };
   return (
     <Box component={Paper} p={2} marginBottom={2}>
       <Typography variant="h5" className={classes.sectionTitle}>
@@ -58,7 +49,7 @@ function PopularList({ items, listTitle }) {
                 <Card className={classes.card}>
                   <CardMedia
                     className={classes.cardMedia}
-                    image={getImageURL(item.backdrop_path)}
+                    image={getImageURL(item.backdrop_path, 780)}
                     title={title}
                   />
                   <CardContent>
@@ -77,9 +68,9 @@ function PopularList({ items, listTitle }) {
   );
 }
 
-PopularList.propTypes = {
+ProgramList.propTypes = {
   items: PropTypes.array.isRequired,
   listTitle: PropTypes.string,
 };
 
-export default PopularList;
+export default ProgramList;
